Connection + Format: Move Validity class and run black
